load legend list dynamically as mainly it is collapsed on first load
