+++ components/map/legend/component.js
@@ -1,32 +1,30 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
+import dynamic from 'next/dynamic';
 
 import classnames from 'classnames';
 
-import debounce from 'lodash/debounce';
-
 import { injectIntl } from 'react-intl';
-import renderHtml from 'html-react-parser';
-
-import {
-  LegendListItem,
-  LegendItemTypes,
-  LegendItemToolbar,
-  LegendItemTimeStep
-} from '~/components/map/legend';
-
-import LegendItemButtonInfo from 'components/map/legend/legend-item-toolbar/legend-item-button-info';
-import LegendItemButtonOpacity from 'components/map/legend/legend-item-toolbar/legend-item-button-opacity';
-import LegendItemButtonVisibility from 'components/map/legend/legend-item-toolbar/legend-item-button-visibility';
 
 import Icon from 'components/ui/icon';
-import Tooltip from 'rc-tooltip';
+import Spinner from 'components/ui/spinner';
 
-import TEMPLATES from './templates';
-import ANALYSIS from './analysis';
-
-import modal from 'services/modal';
-import LayerInfo from 'components/map/layer-info';
+const LegendList = dynamic(
+  () => import('./legend-list'),
+  {
+    ssr: false,
+    loading: () => (
+      <div className="c-legend-list">
+        <div className="c-legend-list__loading">
+          <Spinner
+            isLoading
+            className="-small"
+          />
+        </div>
+      </div>
+    )
+  }
+);
 
 class LegendComponent extends PureComponent {
   static propTypes = {
@@ -37,8 +35,7 @@ class LegendComponent extends PureComponent {
     toolbar: PropTypes.node,
     intl: PropTypes.object.isRequired,
 
-    setLayerSettings: PropTypes.func.isRequired,
-    setLayerOrder: PropTypes.func
+    setLayerSettings: PropTypes.func.isRequired
   }
 
   static defaultProps = {
@@ -53,84 +50,9 @@ class LegendComponent extends PureComponent {
     this.state = { expanded };
   }
 
-  onChangeInfo = (info, id) => {
-    const { layerGroups } = this.props;
-    const layer = layerGroups.find(l => l.id === id);
-
-    modal.toggleModal(true, {
-      children: LayerInfo,
-      childrenProps: {
-        metadata: layer.metadata
-      }
-    });
-  }
-
-  onChangeVisibility = (l, visibility, id) => {
-    const { setLayerSettings } = this.props;
-    setLayerSettings({ id, settings: { visibility } });
-  }
-
-  onChangeOpacity = debounce((l, opacity, id) => {
-    const { setLayerSettings } = this.props;
-    setLayerSettings({ id, settings: { opacity } });
-  }, 250)
-
-  onChangeLayerDate = (dates, layer) => {
-    const { setLayerSettings } = this.props;
-    const { id, decodeConfig } = layer;
-
-    setLayerSettings({
-      id,
-      settings: {
-        ...(decodeConfig && {
-          decodeParams: {
-            startDate: dates[0],
-            endDate: dates[1],
-            trimEndDate: dates[2]
-          }
-        }),
-        ...(!decodeConfig && {
-          params: {
-            startDate: dates[0],
-            endDate: dates[1]
-          }
-        })
-      }
-    });
-  }
-
-  onRemoveLayer = (layer) => {
-    const { toggleLayer } = this.props;
-    toggleLayer(layer);
-  }
-
   onToggleLegend = (bool) => {
     this.setState({ expanded: bool });
   };
-
-  renderDisclaimer = ({ disclaimer, disclaimerTooltip }) => {
-    const { intl } = this.props;
-
-    return intl.formatMessage({ id: disclaimer }, {
-      highlight: chunks => {
-        if (!disclaimerTooltip) return <span className="highlight">{chunks}</span>;
-
-        return (
-          <Tooltip
-            placement="bottom"
-            overlay={
-              <div style={{ maxWidth: 200 }}>
-                {intl.formatMessage({ id: disclaimerTooltip })}
-              </div>
-            }
-            overlayClassName="c-tooltip no-pointer-events"
-          >
-            <span className="highlight">{chunks}</span>
-          </Tooltip>
-        )
-      }
-    });
-  }
 
   render() {
     const { intl, className, collapsable, layerGroups, toolbar, setLayerSettings } = this.props;
@@ -156,66 +78,7 @@ class LegendComponent extends PureComponent {
                 <Icon name="icon-arrow-down" className="-small" />
               </button>
             )}
-
-            {expanded && (
-              <ul className="c-legend-list">
-                {layerGroups.map((layerGroup, i) => (
-                  <LegendListItem
-                    index={i}
-                    key={layerGroup.id}
-                    layerGroup={layerGroup}
-                    {...layerGroup}
-                    toolbar={
-                      toolbar || (
-                        <LegendItemToolbar>
-                          {layerGroup.metadata && <LegendItemButtonInfo />}
-                          <LegendItemButtonOpacity />
-                          <LegendItemButtonVisibility />
-                        </LegendItemToolbar>
-                      )
-                    }
-                    onChangeInfo={(l => this.onChangeInfo(true, layerGroup.id))}
-                    onChangeVisibility={((l, visibility) => this.onChangeVisibility(l, visibility, layerGroup.id))}
-                    onChangeOpacity={(l, opacity) => this.onChangeOpacity(l, opacity, layerGroup.id)}
-                    onRemoveLayer={(l) => { this.onRemoveLayer(l); }}
-                  >
-                    {!!TEMPLATES[layerGroup.id] &&
-                      React.createElement(TEMPLATES[layerGroup.id], {
-                        setLayerSettings
-                      })
-                    }
-
-                    <LegendItemTypes />
-
-                    <LegendItemTimeStep
-                      defaultStyles={{
-                        handleStyle: {
-                          backgroundColor: 'white',
-                          borderRadius: '50%',
-                          boxShadow: '0 1px 2px 0 rgba(0, 0, 0, 0.29)',
-                          border: '0px',
-                          zIndex: 2
-                        },
-                        railStyle: { backgroundColor: '#d6d6d9' },
-                        dotStyle: { visibility: 'hidden', border: '0px' }
-                      }}
-                      handleChange={this.onChangeLayerDate}
-                    />
-
-                    {!!layerGroup.analysis && ANALYSIS[layerGroup.id] &&
-                      React.createElement(ANALYSIS[layerGroup.id], {
-                        analysis: layerGroup.analysis
-                      })
-                    }
-                    {layerGroup.metadata && layerGroup.metadata.disclaimer && (
-                      <div className="legend-item-disclaimer">
-                        {this.renderDisclaimer(layerGroup.metadata)}
-                      </div>
-                    )}
-                  </LegendListItem>
-                ))}
-              </ul>
-            )}
+            {expanded && <LegendList layerGroups={layerGroups} toolbar={toolbar} setLayerSettings={setLayerSettings} />}
           </div>
 
           {/* LEGEND CLOSED */}
